Only try to remove directories

+++ src/launch_revvy.py
@@ -39,10 +39,11 @@ def subprocess_cmd(command):
 
 def cleanup_invalid_installations(directory):
     for fw_dir in os.listdir(directory):
-        fw_dir = os.path.join(directory, fw_dir)
-        manifest_file = os.path.join(fw_dir, 'installed')
-        if not os.path.isfile(manifest_file):
-            shutil.rmtree(fw_dir)
+        if os.path.isdir(fw_dir):
+            fw_dir = os.path.join(directory, fw_dir)
+            manifest_file = os.path.join(fw_dir, 'installed')
+            if not os.path.isfile(manifest_file):
+                shutil.rmtree(fw_dir)
 
 
 def has_update_package(directory):
